PP-30 refactor: move enums at outer file
