+++ pointing-poker/src/enums.ts
@@ -1,0 +1,22 @@
+export enum USER_ROLES {
+  PLAYER = 'player',
+  DEALER = 'dealer',
+  OBSERVER = 'observer',
+}
+
+export enum ROUND_STATUS {
+  NOT_STARTED = 'not started',
+  STARTED = 'started',
+  FINISHED = 'finished',
+}
+
+export enum SCORE_TYPE {
+  fibonacciNumbers = 'fibonacciNumbers',
+  powersOfTwo = 'powersOfTwo',
+}
+
+export enum LOGIN_FORM_INPUTS {
+  FIRST_NAME = 'firstName',
+  LAST_NAME = 'lastName',
+  JOB_POSITION = 'jobPosition',
+}
